Block index, header index and block storage split into in memory / disk/db implementations #334
