+++ src/bitcoin-blockchain/block/index/BlockIndexError.swift
@@ -1,0 +1,3 @@
+enum BlockIndexError: Error {
+    case parentMissing
+}
