+++ src/bitcoin-blockchain/block/storage/BlockStorageError.swift
@@ -1,0 +1,3 @@
+enum BlockStorageError: Error {
+    case dataLocationIssue, missingBlockFiles, blockFileReadIssue, blockFileWriteIssue, blockFileCreateIssue, corruptedBlockData, invalidFileRef
+}
