+++ src/bitcoin-blockchain/block/storage/BlockStorageStatus.swift
@@ -1,0 +1,3 @@
+enum BlockStorageStatus {
+    case idle, starting, running, stopping, stopped
+}
